Focal Point Displays
Added the ability to draw focal points on an image view dynamically.
This entails have points on the image of an original dimension translate
exactly to points on a scaled image.  Currently Rotation is not handled.

Updated the AppTheme Style so that a window overlay to styles App style
sheet and Added A view tool for a horizontal sliding menu